+++ index.tsx
@@ -1,6 +1,7 @@
 // Framework Imports
 import React from 'react'
 import { createRoot } from 'react-dom/client'
+import { BrowserRouter } from 'react-router-dom'
 
 // Component Imports
 import App from './src/App'
@@ -8,4 +9,8 @@ import App from './src/App'
 const container = document.getElementById('root')
 const root = createRoot(container)
 
-root.render(<App />)
+root.render(
+  <BrowserRouter>
+    <App />
+  </BrowserRouter>
+)