+++ src/components/pages/Ottawa/Ottawa.tsx
@@ -1,0 +1,15 @@
+// Framework Imports
+import React, { FC } from 'react'
+
+/**
+ * This the Ottawa component.
+ */
+const Ottawa: FC = () => {
+  return (
+    <section className="ottawaWrap">
+      <h1>Ottawa</h1>
+    </section>
+  )
+}
+
+export default Ottawa